LayoutTests/fast/parser/entity-end-title-tag.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<title id="mytitle">&lt;/title></title>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Test parsing of entity-escaped &lt;/title&gt; tag for <a href=\"https://bugs.webkit.org/show_bug.cgi?id=7931\">Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely</a>");

var expectedResult = '<' + '/title>';
shouldBe('document.getElementById("mytitle").textContent', 'expectedResult');
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
